moved ApplicationPropertiesConstants class
