+++ src/main/java/com/epam/cargo/infrastructure/source/properties/constant/ApplicationPropertiesConstants.java
@@ -1,4 +1,4 @@
-package com.epam.cargo.infrastructure.source.properties;
+package com.epam.cargo.infrastructure.source.properties.constant;
 
 /**
  * Contains all keys of application.properties file.<br>
